Ajout commentaires

+++ my-app/src/components/Log/Logout.js
@@ -1,16 +1,19 @@
 import React from "react";
-
+import {FaUsersSlash} from 'react-icons/fa';
 const Logout = () => {
     const clearToken = () => {
-        localStorage.clear();
-        window.location = '/';
+        if (window.confirm("Voulez-vous vous déconnectez ?")) {
+            localStorage.clear();
+        window.location = '/Connexion';
+        }
+        
     };
 
 
 
     return (
         <li onClick={clearToken}>
-            Se deconnecter
+            <FaUsersSlash/>
         </li>
     );
 };
